ref: extract transofrmations and entities  for db
