+++ packages/entities/indexedDb.ts
@@ -1,0 +1,10 @@
+export interface IConfig {
+  name: string;
+  value: string;
+}
+
+export interface IVotes {
+  id: number;
+  code: number;
+  electionId: string;
+}
